djangoTestServer/requirements.txt
Django==5.2.4
djangorestframework
pymysql
paho-mqtt
opencv-python
numpy
Pillow